RUMM-1599 Create dummy RUM session for crashes occuring in app launch
This corresponds to a situation when the app crashes after initializing Datadog SDK,
but before any RUM view is started. In this case, we don’t have `lastRUMView`
information and we cannot link the RUM error to previous session. Instead, we create
a dummy one, just for sending the crash.